Avoid compiler complaints
